[pull] master from dolphin-emu:master

+++ Source/Core/Core/NetPlayClient.cpp
@@ -2546,7 +2546,7 @@ void NetPlayClient::SendGameStatus()
     }
   }
 
-  packet << static_cast<u32>(result);
+  packet << result;
   Send(packet);
 }
